Some changes, add vercel button, postinstall and guides
