+++ .gitignore
@@ -52,5 +52,4 @@ local-data
 .cursorrules
 .cursor
 *.ignore
-.mcp-config.json
