+++ package.json
@@ -6,7 +6,7 @@
   "license": "MIT",
   "type": "module",
   "scripts": {
-    "dev": "pnpm initial:check && next dev --turbopack",
+    "dev": "next dev --turbopack",
     "build": "next build",
     "start": "next start",
     "build:local": "cross-env NO_HTTPS=true AUTH_TRUST_HOST=true next build",
@@ -15,12 +15,8 @@
     "lint:fix": "next lint --fix && biome lint --write --unsafe",
     "format": "biome format --write",
     "check-types": "tsc --noEmit",
-    "initial:check": "tsx scripts/flag.ts check",
     "initial:env": "tsx scripts/copy-env.ts",
-    "initial:mcp-config": "tsx scripts/generate-mcp-config.ts",
-    "initial:flag-generate": "tsx scripts/flag.ts generate",
-    "postinstall": "tsx scripts/vercelPostinstall.ts",
-    "initial": "pnpm install && pnpm initial:env && pnpm initial:mcp-config && pnpm db:migrate && pnpm initial:flag-generate && pnpm clean .next",
+    "postinstall": "tsx scripts/postinstall.ts",
     "clean": "tsx scripts/clean.ts",
     "db:generate": "drizzle-kit generate",
     "db:reset": "drizzle-kit drop && drizzle-kit push",
