+++ scripts/postinstall.ts
@@ -1,0 +1,43 @@
+import { exec } from "child_process";
+import { IS_VERCEL, IS_DOCKER_BUILD } from "lib/const"; // Import IS_DOCKER as well
+import { promisify } from "util";
+
+const execPromise = promisify(exec);
+
+async function runCommand(command: string, description: string) {
+  console.log(`Starting: ${description}`);
+  try {
+    const { stdout, stderr } = await execPromise(command, {
+      cwd: process.cwd(),
+      env: process.env,
+    });
+
+    console.log(`${description} output:`);
+    console.log(stdout);
+
+    if (stderr) {
+      console.error(`${description} stderr:`);
+      console.error(stderr);
+    }
+    console.log(`${description} finished successfully.`);
+  } catch (error: any) {
+    console.error(`${description} error:`, error);
+    process.exit(1);
+  }
+}
+
+async function main() {
+
+  if (IS_VERCEL) {
+    console.log("Running on Vercel, performing database migration.");
+    await runCommand("pnpm db:migrate", "Database migration");
+  } else if (IS_DOCKER_BUILD) {
+    console.log("Running in Docker, performing database migration.");
+    await runCommand("pnpm db:migrate", "Database migration");
+  } else {
+    console.log("Running in a normal environment, performing database migration and initial environment setup.");
+    await runCommand("pnpm initial:env", "Initial environment setup");
+  }
+}
+
+main();
