+++ src/lib/const.ts
@@ -9,3 +9,4 @@ export const PROMPT_PASTE_MAX_LENGTH = 1000;
 export const IS_VERCEL = process.env.VERCEL === "1";
 
 export const IS_MCP_SERVER_SSE_ONLY = IS_VERCEL;
+export const IS_DOCKER_BUILD = process.env.IS_DOCKER_BUILD === "1"
